Fixed configuration issues and compilation errors
