+++ data/duke.txt
@@ -1,0 +1,2 @@
+T | 0 | ]  haha
+T | 0 | ]  haha /by 1300
