+++ src/main/java/Storage.java
@@ -1,6 +1,10 @@
 import java.io.*;
 import java.nio.file.*;
 import java.util.ArrayList;
+import sam.task.Task;
+import sam.task.Todo;
+import sam.task.Deadline;
+import sam.task.Event;
 
 public class Storage {
     private final Path filePath;
@@ -57,7 +61,7 @@ public class Storage {
     private Task parseTask(String line) {
         // Example: T | 1 | read book
         // Implement parsing logic based on your Task/Deadline/Event/Todo classes
-        String[] parts = line.split(" \| ");
+        String[] parts = line.split(" \\| ");
         try {
             switch (parts[0]) {
                 case "T":
@@ -77,15 +81,15 @@ public class Storage {
     }
 
     private String formatTask(Task task) {
-        String status = task.isDone ? "1" : "0";
+        String status = task.isDone() ? "1" : "0";
         if (task instanceof Todo) {
-            return "T | " + status + " | " + task.description;
+            return "T | " + status + " | " + task.toString().substring(task.toString().indexOf(" ") + 1);
         } else if (task instanceof Deadline) {
             Deadline d = (Deadline) task;
-            return "D | " + status + " | " + d.description + " | " + d.getByRaw();
+            return "D | " + status + " | " + d.toString().substring(d.toString().indexOf(" ") + 1, d.toString().indexOf(" (by:"));
         } else if (task instanceof Event) {
             Event e = (Event) task;
-            return "E | " + status + " | " + e.description + " | " + e.getFromRaw() + " | " + e.getToRaw();
+            return "E | " + status + " | " + e.toString().substring(e.toString().indexOf(" ") + 1, e.toString().indexOf(" (from:"));
         } else {
             return "";
         }
